Small fix to a linking bug. Removed global variable defined in header file
I have no idea how this didn't present itself as a compilation bug
before, but, in my machine, you should not define a variable in a header
file (either .h or .hpp). If you do, the symbol will be created in as
many .o it has been included. This fails to link on my computer. The
linker complains with:

> /usr/bin/ld: .../build/bin/lib/libswm.a(nearest_neighbor_swm_user_code.o):.../src/nearest_neighbor/nearest_neighbor_swm_user_code.h:103: multiple definition of `msg_traffic_def_vector'; libcodes.a(codes-online-comm-wrkld.C.o):.../include/nearest_neighbor_swm_user_code.h:103: first defined here

Notice that this linking error appears only when linking statically. If
we link dynamically, the error never comes up. And that's even worse.
There's a dangling global (non-static) variable which can (or cannot) be
accessed by other processes (?!)

Yeah, better to get rid of such thing by moving it down to where it is
used, the .cpp file.